+++ pkg/translate/convert.go
@@ -1,30 +1,32 @@
-package convert
+package translate
 
 import (
 	"fmt"
 
+	"go.expect.digital/translate/pkg/convert"
 	"go.expect.digital/translate/pkg/model"
 	translatev1 "go.expect.digital/translate/pkg/pb/translate/v1"
 )
 
-func From(schema translatev1.Schema, data []byte) (model.Messages, error) {
+// messagesFromData converts in specific schema serialized data to model.Messages.
+func messagesFromData(schema translatev1.Schema, data []byte) (model.Messages, error) {
 	var from func([]byte) (model.Messages, error)
 
 	switch schema {
 	case translatev1.Schema_ARB:
-		from = FromArb
+		from = convert.FromArb
 	case translatev1.Schema_GO:
-		from = FromGo
+		from = convert.FromGo
 	case translatev1.Schema_JSON_NG_LOCALISE:
-		from = FromNgLocalize
+		from = convert.FromNgLocalize
 	case translatev1.Schema_JSON_NGX_TRANSLATE:
-		from = FromNgxTranslate
+		from = convert.FromNgxTranslate
 	case translatev1.Schema_POT:
-		from = FromPot
+		from = convert.FromPot
 	case translatev1.Schema_XLIFF_2:
-		from = FromXliff2
+		from = convert.FromXliff2
 	case translatev1.Schema_XLIFF_12:
-		from = FromXliff12
+		from = convert.FromXliff12
 	case translatev1.Schema_UNSPECIFIED:
 		return model.Messages{}, fmt.Errorf("unspecified schema")
 	}
@@ -32,24 +34,25 @@ func From(schema translatev1.Schema, data []byte) (model.Messages, error) {
 	return from(data)
 }
 
-func To(schema translatev1.Schema, messages model.Messages) ([]byte, error) {
+// messagesToData converts model.Messages to specific schema serialized data.
+func messagesToData(schema translatev1.Schema, messages model.Messages) ([]byte, error) {
 	var to func(model.Messages) ([]byte, error)
 
 	switch schema {
 	case translatev1.Schema_ARB:
-		to = ToArb
+		to = convert.ToArb
 	case translatev1.Schema_GO:
-		to = ToGo
+		to = convert.ToGo
 	case translatev1.Schema_JSON_NG_LOCALISE:
-		to = ToNgLocalize
+		to = convert.ToNgLocalize
 	case translatev1.Schema_JSON_NGX_TRANSLATE:
-		to = ToNgxTranslate
+		to = convert.ToNgxTranslate
 	case translatev1.Schema_POT:
-		to = ToPot
+		to = convert.ToPot
 	case translatev1.Schema_XLIFF_2:
-		to = ToXliff2
+		to = convert.ToXliff2
 	case translatev1.Schema_XLIFF_12:
-		to = ToXliff12
+		to = convert.ToXliff12
 	case translatev1.Schema_UNSPECIFIED:
 		return nil, fmt.Errorf("unspecified schema")
 	}
